Major Build System Refactor!
- Completely redesigned the flow of the makefile, producing a faster
  makefile experience.

    - Number of conditionals greatly reduced
    - Number of variables greatly reduced
    - Almost all internal variables are marked with a prefix `SJ2_` to
      indicate that they can be read but should not be written to.
    - Number of targets have been greatly reduced to whats needed to
      build, flash and test code and that is all.
    - Simplified the flow of assignments from top to bottom
    - make flag overriding has been removed such as NPROC overriding
    - TESTS and USER_TESTS variables have become 1. Now there is only
      TESTS. To run library tests, add the library test unity_test.cpp
      files into your project.mk file.

- make lint will only check .h and .c files which will relax the amount of
  LINT_FILTER calls needed to ignore 3rd party files. NOTE: all files made
  for SJSU-Dev2 must be .hpp or .cpp in general, but also in order to be
  linted.
- Massive cleanup of almost all sub makefiles. Most of the code in them
  has been removed or consolidated into other files.
- All 3rd Party libraries have been combined into two unity files a .c
  and a .cpp files.
- Moved all continuous integration targets into the new project
  projects/contiuous_integration. Building library tests is now just run
  make test in this folder.
- Made the build process more colorful :)

WARNING documentation has not been updated to reflect these changes.
Documentation will be updated once there has been some testing with the
system and it is proven that this works.

Resolves #1291
Resolves #1292
Resolves #1293

+++ library/L1_Peripheral/test/uart_test.cpp
@@ -118,7 +118,7 @@ TEST_CASE("Testing L1 uart")
           std::array<uint8_t, kPayloadLength> response = {
             0xAA, 0xBB, 0xCC, 0xDD
           };
-          for (int i = 0; i < size; i++)
+          for (size_t i = 0; i < size; i++)
           {
             data[i] = response[read_was_called++];
           }
@@ -160,7 +160,7 @@ TEST_CASE("Testing L1 uart")
           std::array<uint8_t, kPayloadLength> response = {
             0xAA, 0xBB, 0xCC, 0xDD
           };
-          for (int i = 0; i < size; i++)
+          for (size_t i = 0; i < size; i++)
           {
             data[i] = response[read_was_called++];
           }
